move parse tile to cubeclass
